move logic and start of context tracking
